Tiny cleanup in call for glBitmap. Don't see any error in using it for drawing vertices... on OSX this works 100%, also in zbuffer mode and with transparent faces. Code is simple enough for a linuxer to further test!

+++ source/blender/src/glutil.c
@@ -449,7 +449,7 @@ void bglVertex3fv(float *vec)
 	case GL_POINTS:
 		if(pointhack) {
 			glRasterPos3fv(vec);
-			glBitmap(pointhack, pointhack, (float)pointhack/2, pointhack/2, 0.0, 0.0, Squaredot);
+			glBitmap(pointhack, pointhack, (float)pointhack/2.0, (float)pointhack/2.0, 0.0, 0.0, Squaredot);
 		}
 		else glVertex3fv(vec);
 		break;
